pyproject.toml
# Copyright DB InfraGO AG and contributors
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=3.4", "wheel"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]

name = "capella-diff-tools"
description = "Tools for comparing different versions of a Capella model"
readme = "README.md"
requires-python = ">=3.11, <3.14"
license = { text = "Apache-2.0" }
authors = [
  { name = "DB InfraGO AG" },
]
keywords = []
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: Apache Software License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "capellambse>=0.6.6,<0.7",
  "click",
  "diff-match-patch>=20230430",
  "jinja2>=3.1.2",
  "pyyaml>=6.0.1",
]

[project.urls]
Homepage = "https://github.com/DSD-DBS/capella-diff-tools"
Documentation = "https://dsd-dbs.github.io/capella-diff-tools"

[project.optional-dependencies]
docs = [
  "furo",
  "sphinx",
  "sphinx-copybutton",
  "tomli; python_version<'3.11'",
]

test = [
  "pytest",
  "pytest-cov",
]

[project.scripts]
capella-diff-tool = "capella_diff_tools.__main__:main"

[tool.coverage.run]
branch = true
command_line = "-m pytest"
source = ["capella_diff_tools"]

[tool.coverage.report]
exclude_also = [
  'if t\.TYPE_CHECKING:',
  'class .*\bt\.Protocol\):',
  '@abc\.abstractmethod',
  '@t\.overload',
]
skip_covered = true

[tool.docformatter]
wrap-descriptions = 72
wrap-summaries = 79

[tool.isort]
profile = 'black'
line_length = 79

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_unreachable = true
python_version = "3.11"

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_incomplete_defs = false
disallow_untyped_defs = false

[[tool.mypy.overrides]]
# Untyped third party libraries
module = [
  "diff_match_patch.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = """
  --import-mode=importlib
  --strict-config
  --strict-markers
"""
testpaths = ["tests"]
xfail_strict = true

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = [
  "ARG",    # flake8-unused-arguments
  "B",      # flake8-bugbear
  "C4",     # flake8-comprehensions
  "C90",    # mccabe
  "D",      # pydocstyle
  "D212",   # "Multi-line docstring summary should start at the first line"
  "D402",   # "First line should not be the function’s 'signature'"
  "D417",   # "Missing argument descriptions in the docstring"
  "DTZ",    # flake8-datetimez
  "ERA",    # eradicate
  "FA",     # flake8-future-annotations
  "FBT",    # flake8-boolean-trap
  "FIX",    # flake8-fixme
  "FURB",   # refurb
  "G",      # flake8-logging-format
  "ICN",    # flake8-import-conventions
  "ISC001", # "Implicitly concatenated string literals on one line"
  "ISC003", # "Explicitly concatenated string should be implicitly concatenated"
  "LOG",    # flake8-logging
  "PIE",    # flake8-pie
  "PL",     # pylint
  "PT",     # flake8-pytest-style
  "RET",    # flake8-return
  "RUF",    # ruff
  "SIM",    # flake8-simplify
  "TCH005", # "Found empty type-checking block"
  "T1",     # flake8-debugger
  "UP",     # pyupgrade
  "YTT",    # flake8-2020
]
extend-ignore = [
  "D1",      # Missing docstring in _
  "D201",    # No blank lines allowed before function docstring  # auto-formatting
  "D202",    # No blank lines allowed after function docstring  # auto-formatting
  "D203",    # 1 blank line required before class docstring  # auto-formatting
  "D204",    # 1 blank line required after class docstring  # auto-formatting
  "D211",    # No blank lines allowed before class docstring  # auto-formatting
  "D213",    # Multi-line docstring summary should start at the second line
  "DTZ001",  # `tzinfo=None` passed to `datetime.datetime()`
  "DTZ005",  # `tz=None` passed to `datetime.datetime.now()`
  "E402",    # Module level import not at top of file
  "F403",    # `from _ import *` used; unable to detect undefined names
  "F405",    # `_` may be undefined, or defined from star imports
  "PLC0414", # Import alias does not rename original package # used for explicit reexports
  "PLR0904", # Too many public methods
  "PLR0911", # Too many return statements
  "PLR0912", # Too many branches
  "PLR0913", # Too many arguments in function definition
  "PLR0914", # Too many local variables
  "PLR0915", # Too many statements
  "PLR0916", # Too many Boolean expressions
  "PLR0917", # Too many positional arguments
  "SIM108",  # Use ternary operator instead of `if`-`else`-block
]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = [
  "PLE0604", # Invalid object in `__all__`, must contain only strings  # false-positive when unpacking imported submodule __all__
]
"tests/test_*.py" = [
  "F811",    # Redefinition of unused `_` from line _
  "PLR2004", # Magic value used in comparison, consider replacing `_` with a constant variable
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
ignore-decorators = ["typing.overload"]

[tool.ruff.lint.mccabe]
max-complexity = 14

[tool.setuptools]
platforms = ["any"]
zip-safe = false

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.setuptools.packages.find]
include = ["capella_diff_tools", "capella_diff_tools.*"]

[tool.setuptools_scm]
# This section must exist for setuptools_scm to work
local_scheme = "no-local-version"
